New webui submodule system.
This new system is based upon "MetaModules". Each MetaModule contains a
list of modules and can call one or all of them. So the webui doesn't
interract directly with it's modules anymore.

For instance, we now have a AuthMetaModule that call every auth module until
check_auth returns True. This AuthMetaModule even implements a default
behavior when no auth module is declared in the webui.cfg.